openvibe :
 * modified dynamic box context to have both start and end time when sending buffers (start time is no more deduced from last endtime)

openvibe-kernel :
 + auto output chunk reset when marked as ready to send
 + auto output chunk reset before calling notification callbacks and process callback
 + added MaskStart / MaskStop event listeners in order to create / free the OpenViBE box algorithm
 * corrected bug on player deletion so it deletes OpenMASK controller and simtree
 * corrected bug on OpenMASK chunk type so it has 64bits start and end time better than 32 bits (causing rounding errors)
 * implemented dynamic box context to have both start and end time when sending buffers

openvibe-toolkit :
 + added base known openvibe protocol EBML identifiers definition

openvibe-application-designer :
 - removed id-generator (this is a new project)

openvibe-application-acquisition-server :
 - removed EBML node id definition
 + added dependancy to openvibe-toolkit to have node id definitions

openvibe-application-id-generator :
 + created trunc
 * corrected bug when using long long int in place of long int


git-svn-id: svn://scm.gforge.inria.fr/svn/openvibe@182 c330d7e9-fc0c-0410-a5b3-fd85c6f5aa8f

+++ openvibe-applications/id-generator/trunc/src/ovig_main.cpp
@@ -8,8 +8,8 @@ int main(int argc, char** argv)
 	srand(time(NULL));
 	for(int i=0; i<15; i++)
 	{
-		unsigned long long int l_ui32Vale1=(rand()&0x00ffffff);
-		unsigned long long int l_ui32Vale2=(rand()&0xffffffff);
+		unsigned long int l_ui32Vale1=(rand()&0x00ffffff);
+		unsigned long int l_ui32Vale2=(rand()&0xffffffff);
 		printf("#define OV_ClassId_                                        OpenViBE::CIdentifier(0x%08X, 0x%08X)\n", l_ui32Vale1, l_ui32Vale2);
 	}
